+++ framework/Pizza.Mvc/Controllers/AccountController.cs
@@ -7,7 +7,6 @@ using Omu.ValueInjecter;
 using Pizza.Contracts.Security;
 using Pizza.Contracts.Security.ServiceContracts;
 using Pizza.Contracts.Security.ViewModels;
-using Pizza.Framework.ValueInjection;
 using Pizza.Mvc.Security;
 
 namespace Pizza.Mvc.Controllers